Rename Linode utils
