+++ packages/manager/cypress/e2e/core/linodes/linode-config.spec.ts
@@ -12,7 +12,7 @@ import {
 import {
   createLinodeAndGetConfig,
   createAndBootLinode,
-} from 'support/util/linode-utils';
+} from 'support/util/linodes';
 
 import type { Config, Linode } from '@linode/api-v4';
